Full CRUD
Allows the user to create, read, update, and destroy any data in the database. You can add a user, view details of a user and update a user by pressing on the listview, and delete a user with a longclick on the listview. Only problem left is that there is no prompt when you delete a user.